fix(zkp): renamed coin and token files to ft_commitment and nft_commitment respectively
