+++ API-Gateway/src/routes/ft_commitment.js
@@ -1,5 +1,5 @@
 import express from 'express';
-import { mintCoin, transferCoin, burnCoin, checkCorrectnessCoin } from '../services/coin';
+import { mintCoin, transferCoin, burnCoin, checkCorrectnessCoin } from '../services/ft_commitment';
 
 const router = express.Router();
